code/toymcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toymcp"
version = "0.1.0"
description = "A project for learning Model Context Protocol (MCP)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "click>=8.0.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
sse = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=2.0.0",
]

[project.scripts]
toymcp = "toymcp.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/toymcp"]
